+++ packages/ricos-editor/web/src/toolbars/TextFormattingToolbar.tsx
@@ -44,6 +44,10 @@ interface TextFormattingToolbarProps {
 export type TextFormattingToolbarType = typeof TextFormattingToolbar;
 
 class TextFormattingToolbar extends Component<TextFormattingToolbarProps> {
+  updateToolbar = () => {
+    this.forceUpdate();
+  };
+
   render() {
     const {
       activeEditor,